move getDomTarget to a new file with a better name
