+++ packages/base/src/sap/ui/webcomponents/base/DOMEventHandler.js
@@ -1,10 +1,11 @@
 import PseudoEvents from '@ui5/webcomponents-core/dist/sap/ui/events/PseudoEvents';
 import ControlEvents from './events/ControlEvents';
+import getOriginalEventTarget from './events/getOriginalEventTarget';
 
 const handleEvent = function (event) {
 
 	// Get the DOM node where the original event occurred
-	let target = getDomTarget(event);
+	let target = getOriginalEventTarget(event);
 	event.ui5target = target;
 
 	// Traverse the DOM
@@ -17,19 +18,7 @@ const handleEvent = function (event) {
 	}
 };
 
-const getDomTarget = function(event) {
 
-	// Default - composedPath should be used (also covered by polyfill)
-	if (typeof event.composedPath === "function") {
-		const composedPath = event.composedPath();
-		if (Array.isArray(composedPath) && composedPath.length) {
-			return composedPath[0];
-		}
-	}
-
-	// Fallback
-	return event.target;
-};
 
 const processDOMNode = function(node, event) {
 	const id = node.getAttribute("data-sap-ui");
@@ -128,5 +117,4 @@ class DOMEventHandler {
 	}
 }
 
-export { getDomTarget };
 export default DOMEventHandler;
