+++ packages/main/test/sap/ui/webcomponents/main/pages/Eventing.html
@@ -6,6 +6,12 @@
 
     <title>Eventing</title>
 
+    <script data-id="sap-ui-config" type="application/json">
+        {
+            "xx-wc-force-default-gestures": true
+        }
+    </script>
+
     <script src="../../../../../../resources/sap/ui/webcomponents/main/bundle.esm.js"
             type="module"
     >
